move omsecrets package to pkg/omsecrets
